+++ attributes/default.rb
@@ -1,4 +1,4 @@
 default['macosx_autologin']['username'] = nil
 default['macosx_autologin']['password'] = nil
 
-default['macosx_autologin']['restart_loginwindow'] = true
+default['macosx_autologin']['restart_loginwindow'] = false